README.md: Iterate on UX, pkg/cli: add colors, improve help text (#97)
# Description
Propose a change to the landing page of the project to make it more user
friendly

# Context
## README
Changes are fairly self-explanatory. I'm certainly open to feedback on
the structure or syntax. The goal was to approach this with a more "show
-over- tell". So instead of telling people what the value is of the
project, show how they can access it. For the most part, this content
existed before.

## devel/ folder
There was some really useful content in the top level README, but I felt
that it kind of diluted the message in that readme. My proposal here is
to take the content and move it to a dedicated area, and link to it from
the README

## CLI
The CLI now has some color, and an improved help text. 

For posterity, I used https://carbon.now.sh to generate the image in the
README representing the CLI

---------

Signed-off-by: Sam Heilbron <samheilbron@gmail.com>
Signed-off-by: Sam Heilbron <SamHeilbron@gmail.com>
Co-authored-by: Peter Jausovec <peterj@users.noreply.github.com>